Move workspace symbol implementation out of SourceKitLSPServer
The `workspace/symbol`, `workspace/symbolNames`, `workspace/symbolInfo`
and `workspaceSymbol/resolve` implementations made up roughly 470 lines
of SourceKitLSPServer.swift. Move them to WorkspaceSymbols.swift and
leave each request handler in SourceKitLSPServer.swift as a single
statement that forwards to its implementation.

Fold QualifiedWorkspaceSymbolQuery.swift into the same file, since the
query type and its fuzzy matching helper are only used by the qualified
`workspace/symbol` search path.

Also remove `languageClass(for:)`, whose last caller went away when
language service instances became scoped per workspace.
